Small hack to leave content_type untouched.
This hack is largely so one can have application/xml for feeds and
have it set manually while still serving text/html. The problem
being the impedence mismatch between the Accept headers and what's
actually desired. This manifests when one has application/xml in the
config map for feeds and Chrome (and possibly other Webkit browsers)
makes a request with application/xml as the first in the Accept
header list, this has C::A::SerializeBase setting the content_type
to application/xml which isn't desirable when serving text/html.